role, location, room

+++ src/modules/location/v1/location.controllers.ts
@@ -7,15 +7,35 @@ import { PermissionError } from "@/shared/error-handler";
 
 export default class LocationControllers {
   static async getLocations(req: Request, res: Response) {
-    if (!hasPermission(req.roles!, "read:location"))
+    if (!hasPermission(req.roles!, "read:location:*"))
       throw new PermissionError();
 
     const locations = await LocationServices.getLocations();
+
     res.status(StatusCodes.OK).json({
       status: StatusCodes.OK,
       success: true,
-      message: "",
+      message:
+        locations.length == 0
+          ? "Không có kết quả nào"
+          : `Có ${locations.length} giá trị từ yêu cầu.`,
       data: locations,
+    });
+  }
+  static async getLocationById(
+    req: Request<{ locationId: string }>,
+    res: Response
+  ) {
+    const { locationId } = req.params;
+    if (!hasPermission(req.roles!, "read:location:id"))
+      throw new PermissionError();
+    const location = await LocationServices.getLocationById(locationId);
+
+    res.status(StatusCodes.OK).json({
+      status: StatusCodes.OK,
+      success: true,
+      message: "Tìm thấy 1 giá trị",
+      data: location,
     });
   }
 
@@ -23,10 +43,13 @@ export default class LocationControllers {
     req: Request<{}, {}, CreateLocationReq["body"]>,
     res: Response
   ) {
+    if (!hasPermission(req.roles!, "write:location"))
+      throw new PermissionError();
+
     const input = req.body;
     const location = await LocationServices.createLocation(input);
-    res.status(StatusCodes.OK).json({
-      status: StatusCodes.OK,
+    res.status(StatusCodes.CREATED).json({
+      status: StatusCodes.CREATED,
       success: true,
       message: "Tạo địa điểm thành công",
       data: location,
@@ -37,6 +60,9 @@ export default class LocationControllers {
     req: Request<UpdateLocationReq["params"], {}, UpdateLocationReq["body"]>,
     res: Response
   ) {
+    if (!hasPermission(req.roles!, "edit:location"))
+      throw new PermissionError();
+
     const { locationId } = req.params;
     const data = req.body;
 
@@ -57,6 +83,9 @@ export default class LocationControllers {
     req: Request<{ locationId: string }>,
     res: Response
   ) {
+    if (!hasPermission(req.roles!, "delete:location"))
+      throw new PermissionError();
+
     const { locationId } = req.params;
     const location = await LocationServices.deleteLocationById(locationId);
